+++ packages/@sfx/sayt-driver-plugin/CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - The `SaytDriverPlugin` module acts as a mediator between the Logic and View layers for prebuilt modules.
   - SFX-248: The `@sfx/events` package is used for event names and payload interfaces.
   - SFX-333: Added retrieval and event emission of products.
+  - SFX-385: Removed hardcoded product transformer and allowed for passing of a custom transformer.